refactor(mobile): split VoiceInput.tsx 767->256 LOC (Cluster A2)
CLAUDE.md mandates components <400 LOC. VoiceInput.tsx was 767. Extracted the
recording/transcription state machine into a hook, behavior-preserving:
- voice/recording-options.ts: RECORDING_OPTIONS + MAX_RECORDING_MS.
- voice/types.ts: VoiceInputProps + RecordingState.
- voice/useVoiceRecording.ts (356): the full state machine — permission,
  record, stop, transcribe, confirm, pulse animation, auto-stop timer,
  unmount teardown, press handlers. The forward-reference refs
  (stopRecordingRef/transcribeAudioRef) preserve the original TDZ-avoidance
  pattern exactly.
- VoiceInput.tsx (256): the presentational shell (button, recording label,
  confirmation modal, error toast), consuming the hook. Re-exports the public
  types so importers of './VoiceInput' are unchanged.

Tests: existing voice.test.tsx + chat-screen.test.tsx still pass (39 in those
suites), full mobile suite 1715 green = behavior preserved. typecheck + lint
clean; a11y guard passes (all extracted Pressables keep their labels).
